federated_learning/clients/requirements.txt
flwr==1.6.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
mlflow==1.14.0
azureml-sdk